package all debug symbol files in tools-dbg
There is a QA warning that the debug symbol files are installed but not packaged,
if `PACKAGE_DEBUG_SPLIT_STYLE` is set to "debug-file-directory".
Depending on `PACKAGE_DEBUG_SPLIT_STYLE` the debug symbol files are installed in a .debug folder or in /usr/lib/debug.
The sources are in /usr/src/debug
With `PACKAGE_DEBUG_SPLIT_STYLE` set to ".debug" (debug), the debug symbol files are in the efibootguard-tools package
and installed with the binary.

Signed-off-by: Raphael Freudiger <raphael.freudiger@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>