Rename base package to efibootguard
When setting the EFI_PROVIDER variable two things are expected:

1. A package with the name ${EFI_PROVIDER} exists that is included in
packagegroup-core-boot
2. A bbclass with the name ${EFI_PROVIDER} exists

With the current layer there is a class called efibootguard but the package is
called efibootguard-efi. These do not match.

This renames the base package to efibootguard to match the class name.
This is also in line with how systemd-boot and grub-efi work.

Signed-off-by: Pascal Bach <pascal.bach@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>