efibootguard-boot: add a configuration option for kernel name
Instead of hardcoding "bzImage", the efibootguard-boot plugin
is now using a new `kernel` source parameters to select the kernel
file. If the parameter is not set, the default value is now taken
from the KERNEL_IMAGETYPE variable.

To change the filename of the kernel inside the boot partition,
a destination filename can be added to kernel parameter using
the ";" character to separe the source filename from the destination

Usage:
  part --source efibootguard-boot --sourceparams "kernel=bzImage"
  part --source efibootguard-boot --sourceparams "kernel=bzImage;KERNEL"